simplify parameters to a single reserved value and example fix
The large quantity of parameters have been simplified to a single one
given the constraint that only a single parameter must be presented in a
given HCLI Document. Multiple sequential parameter submissions must
happen in sequence across HCLI Documents. This also allows for a
potentially unlimited quantity of user supplied parameters to be
submitted (as opposed to 10, as previously)

The example was also fix to clarify some ambiguities and finer points
with respect to the presence of Option and Commands without a "cli" link
relation to avoid "looping" and having a client select the same Command
or Option repetitively.

_HCLI_version was switched to hcli_version, which looks cleaner and the
same "hcli_" prefix was used for the simple string expansion variable
name in the Parameter description.

Some client side considerations were removed from the Linux/Unix section
and moved to the generic client section.

A few sentences were also corrected to improve clarity throughout the
document.