profile refactor to respect id uniqueness and example adjustment
The profile was adjusted to factor out and reference via fragment the
various redundant ids throughout the profile.

An hcli-id document was created to replace treating the entire profile
as one large descriptor and the notional "HCLI Document".

hcli-document is now a fragmented out semantic type of its own and that
can be referenced as such in the type link relation.

various command line examples have been updated to reflect pointing to
the hcli-document fragment in their type link relation.