move to support fragment id reference in profiles
To remain HAL standard and to provide a hint to clients on what to
expect wrt to the expected target semantics of a link or form, the spec
now allows specifically referencing a descriptor.id in a profile link or
profile attribute.

This will help clients make decisions based on the expected target
semantics they can see on links or forms, without them having to specifically
navigate into target resources to discover more specific ALPS semantic information.